move pregnancy fields from model to model
